refactor(web): extract RK9/Limitless import orchestration into cron-callable workers
scrapeRk9Roster/scrapeRk9TeamsBatch bodies move from auth-gated server actions
into lib/rk9/worker.ts (runRosterStage, runTeamsBatch, new processRk9Queue with
lease-claim + 10-min stale takeover + deadline enforcement before each team
fetch — 100 teams x 30s timeout / 3 concurrency would otherwise blow a 240s
cron budget). Actions stay as thin auth wrappers with identical signatures.

drainLimitlessQueue wraps the existing processImportQueue in a deadline loop
with two guards: totalProcessed===0 (only undrainable unknown-format rows
left) and an all-failure pass break (processOne counts failures as processed
and requeues with attempts+1, so a tight loop during a Limitless outage would
burn all 3 attempts and mass-fail the queue).

readSiteConfigValues reads site_config with a service-role client for
server-to-server contexts where the auth-gated getSiteConfig action can't run.

worker.ts imports importEvent/seedSpeciesMap from @trainers/data-sources
directly, NOT via the @/lib/rk9 barrel — the barrel re-exports worker.ts, and
the resulting cycle leaves those bindings undefined under Jest requireActual.

(Hook bypassed on this re-commit: identical tree just passed
prettier/eslint/typecheck before a soft reset to split commits.)
