+++ apps/web/src/lib/limitless/index.ts
@@ -1,3 +1,12 @@
 export * from "./api";
-export type { QueueProcessResult, BatchQueueResult } from "@trainers/data-sources";
-export { syncTournamentList, importTournament, processImportQueue } from "@trainers/data-sources";
+export type {
+  QueueProcessResult,
+  BatchQueueResult,
+} from "@trainers/data-sources";
+export {
+  syncTournamentList,
+  importTournament,
+  processImportQueue,
+} from "@trainers/data-sources";
+export { drainLimitlessQueue } from "./queue-worker";
+export type { DrainResult } from "./queue-worker";
